Let to change the chrome provider for locales
Experimental, enable it setting the preference:
extensions.simplels.provider

...to the name of any locale provider registered via chrome.manifest

It can be useful to developers testing the translation of their own
UI components in any locale supported by *their* extension.

Note that the "User Interface Language" window always will pick up
the list of available locales from the global provider, for to prevent
confusion from the text descriptions there, so it leaves, for the
moment, only to the toolbar button taking advantage of it.