Fix rare issue using a provider from a restartless extension
It's known issue that a restartless disabling of the addon providing
the package used as locale provider is not detected by the extension.
It seems a pain fixing it, for now it is better documented.

A second, deeper, related issue is being fixed here: the menu of the
toolbar button can show no available locales for a custom provider
after triggering the former issue. The full steps for reproducing it
are:

* Set a package PK from a restartless extension EXT as locale provider
* Disable EXT (The menu isn't updated: it's issue 1)
* Install or disable a language pack (The menu is updated, no
  available locales now: an effect of issue 1)
* Reset the provider preference. All should be ok now
* Re-enable EXT
* Set again PK as provider. Now the menu shows no available locales:
  wrong