Leave the community-supported localizations unregistered
It allows anybody to run the latest development snapshot of the addon
simply adding all the files to a zip-renamed xpi file and dragging it
to a window of the application. It, at least, until that the build
tools are added to the repo.

Registration of these locales will happen only as part of the releasing
process.