Add icon to the "Learn more about langpacks" button
Linux only.

It matches the icon settings of other native items linking to SUMO,
as the "Help" item in the main Help menu, or the Help button in the
main options window of Firefox.

To see it requires having enabled the respective GTK setting to show
icons in dialog buttons ("gtk-button-images").