Remove usage of FUEL/STEEL/SMILE library
The FUEL library has been marked as deprecated since Fx 40:
  https://bugzilla.mozilla.org/show_bug.cgi?id=989307

This change re-implement internally Application.restart(), as it's
our only dependence of FUEL in Firefox. It removes our dependence
of STEEL/SMILE in Thunderbird & SeaMonkey too.

The new restartApplication() method in the BrowserUtils.jsm module
isn't used directly (with the proper fallbacks) as it landed broken in
Fx 40 (fixed in Fx 41):
  https://bugzilla.mozilla.org/show_bug.cgi?id=1185294