Support in the most basic way to use flags for locales
First, I have to recall that it's not a good practice to use country
flags to identify languages:
  https://www.cs.tut.fi/~jkorpela/flags.html

So this add-on will not have in the future any exposed option related
to that idea, as I don't want to encourage it.

This patch provides the minimum fundations to allow to users to have
images (e.g. flags) instead of language codes in the menu of the
toolbar button (if the option to show them is enabled) in Thunderbird,
SeaMonkey and old Firefox releases (until Fx 28). Recent Firefox
releases with Australis UI aren't supported yet as the display of the
lang codes is implemented there in a totally different way. The image
files must be provided by the user himself and set manually, e.g. via
userChrome.css with the following code:

  #simplels-button[descriptions="true"] .simplels-locale {
    -moz-binding: url("chrome://simplels/content/menu.xml#menuitem-iconic-locale-flag") !important;
  }
  .simplels-locale-flag[locale="en-US"] { list-style-image: url("flags/en-US.png"); }
  .simplels-locale-flag[locale="de"]    { list-style-image: url("flags/de.png"); }

...and so on, one extra line for each locale, and having the mentioned
flags/ folder located next to the userChrome.css file.

How to support it in current Firefox releases is pending of research.