Listen again the removal of langpacks for Gecko 47+
Gecko 47 changed which events are broadcasted when the availability of
a langpack changes:
  https://bugzilla.mozilla.org/show_bug.cgi?id=612168

Now the onDisabled and onEnabled events aren't triggered when an add-on
is removed, or that removal is undone. Due to this, the menu of our
toolbar button wasn't being updated, by example, immediately after the
removal of a language pack, showing then some inconsistent data.