Link to the proper folder for langpacks in beta builds
Until some time ago, it wasn't possible to know, for this add-on, the
exact version number of the application in beta builds, due to
Mozilla treating each beta build as a "candidate release build" (sort of),
so the beta number was not exposed. It changed:

https://bugzilla.mozilla.org/show_bug.cgi?id=1145175

...so, at last, we can link to the proper container folder for langpacks
in the Mozilla ftp repository.

The latest release 2.48b1 of SeaMonkey still don't use it, so it's not
changed for SM for now.