Update link to langpacks for Firefox Dev Edition
The so-called "Developer Edition" of Firefox, based on Aurora builds,
has been changed to be based now on a sort of special build of Beta,
as the Aurora update channel will be removed:

  https://hacks.mozilla.org/2017/04/simplifying-firefox-release-channels/
  https://bugzilla.mozilla.org/show_bug.cgi?id=1353821

This commit updates the link to the container folder of langpacks in the
Moz archive repository.

The initial new DevEdition builds released recently still use "aurora"
as the value for the "app.update.channel" preference. I don't know yet
if it will change, so, for now, no changes are required to how the
application build is detected by the add-on.

Also, the same change for SeaMonkey was already done: no changes are
required actually. The case of Thunderbird is pending:

  https://bugzilla.mozilla.org/show_bug.cgi?id=1353924